Issue #441
Added code to ensure that the final `run` step takes place when using `File` with Windows or Mac hosts. Previously, the `run` step was never reached because the `extract` step (only necessary when using the `Archive` and `Run` parameters) was not processed. Now that step will be effectively ignored when using the `File` parameter for Mac and Windows which should lead everything completing successfully.