Update devices.ps1
Reduced group submission size for `Remove-FalconHostGroup` to reduce `500: Contact Support` errors. Despite deleting Host Groups when 50 are submitted at a time, there seems to be a server that is taking too long to issue a response which causes the 500 error. Submitting smaller groups avoids the error and gives are proper response back.