Issue #450
Updated `Invoke-TagScript` function to properly attempt Real-time Response script execution when user tries to delete the only tag that is present on target host.

Also moved code higher in function to avoid using Real-time Response script when tag to remove is not present, or when tag to add is already present.